abyan_project_automation/requirements.txt
Appium-Python-Client==3.1.0
pytest==7.4.0
selenium==4.15.1
faker
pytest-html